+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Default emulators:start to use fast dev-mode for Nuxt3 applications (#5551)
 - Fix broken Functions CLI experience for projects with incomplete GCF 2nd Gen functions (#5684)
+- Disable GCF breaking change to automatically run npm build scripts as part of function deploy (#5687)